Work on captcha screen Dashboard
