+++ app/client/components/dashboard/Locked.jsx
@@ -1,0 +1,98 @@
+import { StyleSheet, Text, View, Image } from 'react-native';
+import { useState } from 'react';
+import { PrevButton, GreenButton } from '../navigation/Buttons';
+
+const Locked = ({ onUnlocked }) => {
+  const [showQuestion, setShowQuestion] = useState(true);
+  const [showResult, setShowResult] = useState(false);
+
+  const handleAnswerClick = (answer) => {
+    if (answer === 11) {
+      onUnlocked();
+    } else {
+      // Display "Incorrect, try again" message and reset selected answer
+      setShowQuestion(false);
+      setShowResult(true);
+    }
+  };
+
+  const handleTryAgain = () => {
+    // Reset the state to show the question and hide the result
+    setShowQuestion(true);
+    setShowResult(false);
+  };
+
+  return (
+    <View className='flex flex-column'>
+      <View className='flex p-4 ml-0 mr-auto '>
+        <PrevButton />
+      </View>
+      <View className='container flex items-center justify-center m-auto mt-20 sm:py-24 w-5/12 shadow-lg px-5 py-10 rounded-lg border border-beige bg-[#fdfdf3dd]'>
+        {showQuestion && (
+          <>
+            <Text
+              className="flex text-primary-dark text-3xl my-5 font-black font-['Oleo Script']"
+              style={styles.title}
+            >
+              How much is 5 + 6 ?
+            </Text>
+            <View className='flex flex-row'>
+              <GreenButton onPress={() => handleAnswerClick(11)} text={'11'} />
+              <GreenButton onPress={() => handleAnswerClick(18)} text={'18'} />
+              <GreenButton onPress={() => handleAnswerClick(14)} text={'14'} />
+            </View>
+          </>
+        )}
+        {showResult && (
+          <Text>
+            {' '}
+            Incorrect,{' '}
+            <button onClick={handleTryAgain}>
+              <Text>Try Again</Text>
+            </button>
+          </Text>
+        )}
+      </View>
+    </View>
+  );
+};
+
+export default Locked;
+
+const styles = StyleSheet.create({
+  flex: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  end: {
+    justifyContent: 'flex-end',
+    flexDirection: 'row',
+  },
+  title: {
+    fontWeight: 'bold',
+  },
+  subtitle: {
+    lineHeight: '1.5em',
+    fontSize: '1.125rem',
+    marginVertical: '1em',
+    textAlign: 'center',
+  },
+  text: {
+    lineHeight: '1.5em',
+    fontSize: '1.125rem',
+    marginVertical: '1em',
+    width: '120px',
+    marginRight: '1em',
+  },
+  link: {
+    color: '#1977f2',
+  },
+  listitem: {
+    marginVertical: '0.5rem',
+  },
+  pageLink: {
+    fontSize: '1.25rem',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+});
